feat: 添加接口分析 future: 完善网关，分析excel表数据

+++ bing-api-common/src/main/java/com/abing/model/domain/UserInterfaceInfo.java
@@ -32,7 +32,7 @@ public class UserInterfaceInfo implements Serializable {
     /**
      * 总调用次数
      */
-    private Integer totalNum;
+    private Long totalNum;
 
     /**
      * 剩余调用次数
